Create EMA 09 20 50 100 200
Exponential Moving Averages (EMA) with periods of 9, 20, 50, 100, and 200 on a TradingView chart:
This script will plot the EMAs of the close price of the chart with the corresponding color and linewidth. It will also add a legend on the top of the chart.

To deploy this script on TradingView, you'll need to have an account on the platform. Once you're logged in, follow these steps:

Click on the "New Script" button in the top right corner of the platform.
Copy and paste the script into the editor.
Click on the "Save" button in the top right corner of the script editor.
You can customize your script as per your requirement by using different indicators and modification to the script
You can then add the script to a chart by clicking the "Add Indicator" button on the chart and selecting the script from the list of custom indicators.
Alternatively, you can also share your script by clicking on the three dots on the top right corner of the script editor, and selecting the "Share" option. This will generate a link that you can share with others, so they can also add your script to their charts.

Please keep in mind that it's always good to test script in the strategy tester before using it in live trading and also check how it behaves in different market conditions.