Create Triangle Pattern
This script defines a function called triangle that takes two arguments: trend_a and trend_b. These arguments are used to create two trendlines using the trendline.new() function, which takes a value for the angle of the trendline and a name for the trendline as arguments. The script then checks if the two trendlines intersect using the intersects() method, which is called on one of the trendlines and takes the other trendline as an argument. If the trendlines do intersect, the script uses the plotshape() function to draw a triangle on the chart.

To deploy the script, you will need to copy and paste the code into the PineScript editor within TradingView. Once the script is pasted in the editor, you can edit the angle of the trendline as per your requirement and run the script by clicking the "Apply" button.

It's important to note that for this script to function correctly, the chart must have at least 70-110 bars to form a triangle shape. Also this script is just an example and you should backtest and optimize it based on your preference and the market you are trading.

In addition to this you can add other features and logic in the script like alerts, popups, or sound etc. you can also change the shape of the triangle and its properties like color, size, location etc.