Create Cup and Handle Pattern
This script defines a function called cup_and_handle that takes a single argument length, which is used to specify the number of bars over which the pattern is determined.

The function first uses the highest() and lowest() functions to get the highest high and lowest low within the given length. Then it plots two points on the chart to mark the highest high and lowest low which make the "cup" shape.

Next, the script uses the height of the cup to calculate the handle height and plots two points for the handle top and bottom using the handle_top and handle_bottom calculations.

It also plots a line breakout which is the highest high + (handle_height * 0.382) this line will act as resistance.

Finally, the script uses an if statement to check if the closing price of the current bar is greater than the breakout line, If this is true, it plots an upward arrow on the chart to indicate a potential buying opportunity.

To deploy the script in trading view, You can copy the script and paste it into the PineScript editor within TradingView, then run the script by clicking the "Apply" button. You can change the input parameter to your desired period or as per your requirement.

You can also customize the script by adjusting the colors and styles of the plots, or by adding additional features like alerts or pop-