Create Falling Wedge
In this script, the study function creates the indicator, with the "Falling Wedge" title and the overlay option set to true. Two trendlines are defined, one for the high prices and one for the low prices, with an angle of -45 and 45 degrees respectively, and both trendlines are set to red. The fill function creates the wedge shape by connecting the two trendlines and filling it with a semi-transparent red color. The last line checks if the current bar is within the wedge by checking if the trendlines are crossed, if true the script will alert a message on the top of the chart.

To deploy the script, you will need to access the TradingView platform and open a chart for the asset you want to apply the indicator to. From there, you can open the Pine Editor by clicking on the "Pine Editor" button on the top right corner of the chart. In the editor, you can copy and paste the script and then click the "Save" button. After that, the indicator will appear in the "Studies" section of the chart, and you can apply it by clicking on it. Also you can also use the published script to any other chart by clicking on the import button on Pine Editor.

Please note that, tradingview have some specific policies on usage of Alerts, do refer them before using the script in live trading.