This script uses the hlines() function to plot horizontal lines at the highest high and lowest low over a specified period, length, which can be adjusted to suit your trading strategy. These trend lines form the Pennant or Flag pattern.
The plotshape() function is used to plot the Pennant or Flag pattern on the chart as a triangle. The label.new() function adds a label to the chart showing the bar number on which the pattern was identified.

To deploy this script in TradingView, you will need to first open a chart of the asset you want to analyze. Then, open the Pine Editor by clicking on the Pine Editor button on the chart's toolbar. In the Pine Editor, you can copy and paste the script and then click on the Save button to run the script on the chart.

Output will be a triangle shape showing when flag pattern is formed. Also a label will be appear on chart showing the bar number on which flag pattern was identified.