This script uses the sma() function to calculate the simple moving average of the closing prices over a user-defined lookback period (which can be set using the input() function). The script then checks if the current moving average is greater than the moving average from the previous three bars and plots a green upward-pointing triangle if the criteria are met.
To deploy this script in TradingView, you will need to create a new indicator on the platform and copy and paste the code into the editor. After that you will need to run the backtest and adjust the parameters as needed.

Finally, the output of this script will be a green upward-pointing triangle on the chart, whenever the script identifies the condition met. It will help in identifying the possible Rounding Bottom pattern on the chart and assist in taking trading decision.