Create insidebar_rsi.pine
To deploy this strategy in TradingView, you will need to create a new chart and add the Pine Script code to it. Here are the steps to do this:

Go to TradingView and create a new chart by selecting the asset you want to trade.

Click on the "Pine Editor" button on the bottom of the chart to open the Pine Script editor.

Paste the above code into the editor.

Click on the "Save" button to save the script.

Click on the "Apply" button to apply the script to the chart.

Adjust the parameters of the script to your liking.

Use the buy and sell signals generated by the strategy to make your trading decisions.

It's important to keep in mind that this is a simplified example and it's recommended to test and evaluate the performance of the strategy on historical data before using it for live trading. In addition, it's also important to consider the potential risks and not to rely solely on this strategy to make all your trading decisions.